requirements.txt
websockets==8.1
Flask==1.1.2
mysql-connector-python==8.3.0
requests==2.27.1
numpy==1.24.3